添加了基础弹幕View的碰撞函数

+++ app/src/main/java/com/example/songye02/diasigame/utils/CollisionUtil.java
@@ -5,7 +5,7 @@ package com.example.songye02.diasigame.utils;
  */
 
 public class CollisionUtil {
-    public static boolean isCollisonWithRect(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2) {
+    public static boolean isCollisonWithRect(float x1, float y1, float w1, float h1, float x2, float y2, float w2, float h2) {
         if(x1>=x2+w2){
             return false;
         }else if(x1+w1<=x2){
